@@ -1,10 +1,9 @@
 ﻿using System.Collections.Generic;
-using Microsoft.OpenApi.Models;
+using Microsoft.OpenApi;
 
-namespace Swashbuckle.DocumentTags.Extension
+namespace Swashbuckle.DocumentTags.Extension;
+
+public class DocumentTagsConfig
 {
-    public class DocumentTagsConfig
-    {
-        public IEnumerable<OpenApiTag> Tags { get; set; }
-    }
+    public IEnumerable<OpenApiTag> Tags { get; set; } = [];
 }
